+++ customer_product_qrcode/__manifest__.py
@@ -24,7 +24,8 @@
     'name': 'Customer and Product QR Code Generator',
     'version': '15.0.1.0.0',
     'summary': 'Generate Unique QR Codes for Customers and Products',
-    'description': 'QR Code, QR Code Generator, Odoo QR Code Generator, Customer QR Code, Product QR Code, QR, QR Code Odoo',
+    'description': 'QR Code, QR Code Generator, Odoo QR Code Generator, '
+                   'Customer QR Code, Product QR Code, QR, QR Code Odoo',
     'category': 'Extra Tools',
     'author': 'Cybrosys Techno solutions',
     'maintainer': 'Cybrosys Techno Solutions',
@@ -32,9 +33,13 @@
     'website': 'https://www.cybrosys.com',
     'depends': ['base', 'sale', 'stock'],
     'data': [
+        'data/ir_sequence_data.xml',
         'report/paperformat.xml',
         'report/report.xml',
-        'views/view.xml',
+        'views/product_product_views.xml',
+        'views/product_template_views.xml',
+        'views/res_config_settings_views.xml',
+        'views/res_partner_views.xml',
         'report/template.xml',
     ],
     'images': ['static/description/banner.png'],
