+++ customer_product_qrcode/data/ir_sequence_data.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <data noupdate="1">
+        <record id="ir_sequence_customer" model="ir.sequence">
+            <field name="name">customer_sequence</field>
+            <field name="code">res.partner</field>
+            <field name="prefix"></field>
+            <field name="padding">5</field>
+        </record>
+    </data>
+    <data noupdate="1">
+        <record id="ir_sequence_product" model="ir.sequence">
+            <field name="name">product_sequence</field>
+            <field name="code">product.product</field>
+            <field name="prefix"></field>
+            <field name="padding">5</field>
+        </record>
+    </data>
+</odoo>
