+++ customer_product_qrcode/doc/RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ## Module <customer_product_qrcode>
 
-#### 18.11.2021
-#### Version 15.0.1.0.0
-#### ADD
-- Initial commit for customer_product_qrcode
+#### 10.07.2023
+#### Version 15.0.2.0.0
+#### UPDT
+
+-  Code Refactored
