+++ customer_product_qrcode/views/product_product_views.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!--  Record for inherit Product form to add sequence and generate qr button  -->
+    <record id="product_form_inherit" model="ir.ui.view">
+        <field name="name">product.product.form.qr.inherit</field>
+        <field name="model">product.product</field>
+        <field name="inherit_id" ref="product.product_normal_form_view"/>
+        <field name="arch" type="xml">
+            <div name="button_box" position="inside">
+                <button name="generate_qr" type="object" class="btn-box"
+                        icon="fa-qrcode">
+                    <field name="sequence" invisible="1"/>
+                    Generate QR
+                </button>
+            </div>
+            <field name="categ_id" position="after">
+                <field name="sequence"/>
+            </field>
+        </field>
+    </record>
+</odoo>
