+++ customer_product_qrcode/views/product_template_views.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!--  Record for inherit Product template form to add generate qr button  -->
+    <record id="product_template_form_inherit" model="ir.ui.view">
+        <field name="name">product.template.form.qr.inherit</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id"
+               ref="product.product_template_only_form_view"/>
+        <field name="arch" type="xml">
+            <div name="button_box" position="inside">
+                <button name="generate_qr" type="object" class="btn-box"
+                        icon="fa-qrcode">Generate QR
+                </button>
+            </div>
+        </field>
+    </record>
+</odoo>
