+++ customer_product_qrcode/views/res_config_settings_views.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!--  Record for inherit ResConfigSettings view to add prefix fields in general settings  -->
+    <record id="settings_form_inherit_for_qr" model="ir.ui.view">
+        <field name="name">res.config.inherit.qr</field>
+        <field name="model">res.config.settings</field>
+        <field name="inherit_id"
+               ref="base_setup.res_config_settings_view_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//div[@id='companies']" position="after">
+                <div id="setup_qrcode">
+                    <h2>Setup QRCode</h2>
+                    <div class="row mt16 o_settings_container">
+                        <div class="col-xs-12 col-md-6 o_setting_box">
+                            <div class="o_setting_right_pane">
+                                <label string="Prefixes"
+                                       for="customer_prefix"/>
+                                <span class="fa fa-lg"/>
+                                <div class="text-muted">
+                                    Set your unique prefix.
+                                </div>
+                                <div class="content-group">
+                                    <div class="mt16 row">
+                                        <label for="customer_prefix"
+                                               class="col-xs-3 col-md-6 o_light_label"/>
+                                        <field name="customer_prefix"
+                                               class="oe_inline"/>
+                                        <label for="product_prefix"
+                                               class="col-xs-3 col-md-6 o_light_label"/>
+                                        <field name="product_prefix"
+                                               class="oe_inline"/>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </xpath>
+        </field>
+    </record>
+</odoo>
