+++ customer_product_qrcode/views/res_partner_views.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <!--  Record for inherit Partner form to add sequence and generate qr button  -->
+    <record id="partner_form_inherit" model="ir.ui.view">
+        <field name="name">res.partner.form.qr.inherit</field>
+        <field name="model">res.partner</field>
+        <field name="inherit_id" ref="base.view_partner_form"/>
+        <field name="arch" type="xml">
+            <div name="button_box" position="inside">
+                <button name="generate_qr" type="object" class="btn-box"
+                        icon="fa-qrcode">
+                    <field name="sequence" invisible="1"/>
+                    Generate QR
+                </button>
+            </div>
+            <field name="category_id" position="after">
+                <field name="sequence"/>
+            </field>
+        </field>
+    </record>
+</odoo>
